+++ app/src/test/java/woowacourse/payments/domain/CardNumberTest.kt
@@ -1,12 +1,11 @@
-package woowacourse.payments
+package woowacourse.payments.domain
 
 import io.kotest.assertions.throwables.shouldThrow
 import org.junit.jupiter.api.Test
-import woowacourse.payments.domain.CardNumber
 
 class CardNumberTest {
     @Test
-    fun 카드_번호는_문자가_포함되지_않는다() {
+    fun `카드_번호는_문자가_포함되지_않는다`() {
         // given
         val numbers = "abcd1234test5678"
 
@@ -17,7 +16,7 @@ class CardNumberTest {
     }
 
     @Test
-    fun 카드_번호는_16자리_이하여야_한다() {
+    fun `카드_번호는_16자리_이하여야_한다`() {
         // given
         val numbers = "11112222333344445555"
